Update to new windowManager and game state changes
Made changes to make the general game state and window engine cleaner -
AKA, backporting changes from Lumina. Removed the switching game state
entirely - it was set when the player's were switching positions, but
was entirely pointless (was never conditionally checked and just makes
for multiple states the game can be in while playing normally). Made
pausing into a game state instead of a separate variable. Updated to
newer version of the windowManager engine containing new functionality,
and made appropriate changes to the UI declarations as per changes to
the window engine.